project-fixed/microservices-app/backend/order-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
